requirements.txt
Django>=4.2
djangorestframework>=3.14
django-allauth>=0.60
django-cors-headers>=4.0
pandas>=2.0
openpyxl>=3.1
reportlab>=4.0
